♻️ refactor: 카테고리 페이지
